Agente_IA/requirements.txt
langchain-mcp-adapters
langchain-openai
langgraph
python-dotenv